added favicon and delete book function

+++ public/js/admin.js
@@ -1,20 +1,21 @@
-const deleteProduct = document.querySelector(".delete-product-btn");
+document.addEventListener("click", async function (e) {
+  if (e.target && e.target.classList.contains("delete-product-btn")) {
+    const deleteProduct = e.target;
+    const bookId =
+      deleteProduct.parentNode.querySelector("[name=bookId]").value;
+    const csrfToken =
+      deleteProduct.parentNode.querySelector("[name=_csrf]").value;
 
-deleteProduct?.addEventListener("click", async function (e) {
-  const bookId = this.parentNode.querySelector("[name=bookId]").value;
-  const csrfToken = this.parentNode.querySelector("[name=_csrf]").value;
+    const card = deleteProduct.closest("article");
 
-  const card = this.closest("article");
+    const data = await fetch(`/admin/delete-book/${bookId}`, {
+      method: "DELETE",
+      headers: {
+        "csrf-token": csrfToken,
+      },
+    });
 
-  const data = await fetch(`/admin/delete-book/${bookId}`, {
-    method: "DELETE",
-    headers: {
-      "csrf-token": csrfToken,
-    },
-  });
-
-  const response = await data.json();
-  //   card.remove();
-  card.parentNode.removeChild(card);
-  //   console.log("clicked");
+    const response = await data.json();
+    card.parentNode.removeChild(card);
+  }
 });
